requirements.txt
# Core dependencies
pybind11>=2.10.0
numpy>=1.21.0

# Visualization
pygame>=2.5.0
pillow>=9.0.0  # For GIF recording

# Development dependencies
pytest>=7.0.0
pytest-cov>=3.0.0

# Build dependencies
cmake>=3.12
